requirements.txt
praw
transformers
torch
flask
python-dotenv
